ci: ensure all the images we built are signed with cosign
Prior to this commit, we did NOT sign the container images built by our
pipelines. We only signed the artifacts attached to the final GitHub
release (like the SBOM and the provenance file).

With this commit we sign (using both cosign v2 and v3 formats):
- architecture specific images
- multi-arc image manifest

Currently we support only amd64 images.

Co-authored-by: Flavio Castelli <fcastelli@suse.com>
Signed-off-by: Sam Wang (holyspectral) <sam.wang@suse.com>